some stuff and some don't. it's android studio go figure

+++ app/src/main/java/com/Facybook_android/app/API/PostAPI.java
@@ -146,12 +146,14 @@
             call.enqueue(new Callback<List<Post>>() {
                 @Override
                 public void onResponse(Call<List<Post>> call, Response<List<Post>> response) {
-                    new Thread(() -> {
-                        dao.deleteAll();
-                        dao.insert(response.body());
-                        postListData.postValue(response.body());
-                        Log.e("UserRepository", "Fetched friends' posts successfully");
-                    }).start();
+                    if (response.isSuccessful()) {
+                        new Thread(() -> {
+                            dao.deleteAll();
+                            dao.insert(response.body());
+                            postListData.postValue(response.body());
+                            Log.e("UserRepository", "Fetched friends' posts successfully");
+                        }).start();
+                    }
                 }
                 @Override
                 public void onFailure(Call<List<Post>> call, Throwable t) {
